@@ -5,14 +5,14 @@
     "build": "tsc --build ./tsconfig.json",
     "build.watch": "tsc --build --watch ./tsconfig.json",
     "schedule-workflow": "node ./lib/starter",
-    "start": "node ./lib/worker"
+    "worker": "node ./lib/worker"
   },
   "name": "temporal-hello-world",
   "dependencies": {
-    "temporalio": "0.3.x"
+    "temporalio": "0.4.x"
   },
   "devDependencies": {
-    "typescript": "^4.2.2",
+    "typescript": "^4.4.2",
     "@tsconfig/node14": "^1.0.0"
   }
 }
