@@ -1,14 +1,13 @@
-'use strict';
-
-const { Connection, WorkflowClient } = require('@temporalio/client');
-const uuid = require('uuid');
+import { Connection, WorkflowClient } from '@temporalio/client';
+import { Expense } from '../interfaces/workflows';
+import { v4 } from 'uuid';
 
 async function run() {
   const connection = new Connection();
   const client = new WorkflowClient(connection.service);
-  const expense = client.stub('expense', { taskQueue: 'tutorial_expense1' });
+  const expense = client.stub<Expense>('expense', { taskQueue: 'tutorial' });
 
-  const expenseId = uuid.v4();
+  const expenseId = v4();
 
   await expense.start(expenseId);
 
