@@ -1,16 +1,16 @@
-'use strict';
-
-const { Connection, WorkflowClient } = require('@temporalio/client');
+import { Connection, WorkflowClient } from '@temporalio/client';
+import { Progress } from '../interfaces/workflows';
 
 async function run() {
   const connection = new Connection();
   const client = new WorkflowClient(connection.service);
 
-  const progress = client.stub('progress', { taskQueue: 'tutorial' });
+  const progress = client.stub<Progress>('progress', { taskQueue: 'tutorial' });
 
   await progress.start();
 
   const val = await progress.query.getProgress();
+  // Should print "10", may print another number depending on timing
   console.log(val);
 
   const result = await progress.result();
